Fix component tree
